Update from releng/13.2, git revision 08b87f63a04.

+++ sys/dev/hwpmc/hwpmc_uncore.h
@@ -25,7 +25,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $FreeBSD: ac1623e7737ee53f6f5d13cf7e3cae100e878a28 $
+ * $FreeBSD: 5e8610f77ddf7371d4fb94ffedfacf7c4abf85dc $
  */
 
 #ifndef _DEV_HWPMC_UNCORE_H_
@@ -45,7 +45,7 @@ struct pmc_md_ucf_op_pmcallocate {
  * Programmable PMCs.
  */
 struct pmc_md_ucp_op_pmcallocate {
-	uint32_t	pm_ucp_config;
+	uint64_t	pm_ucp_config;
 };
 
 #define	UCP_EVSEL(C)	((C) & 0xFF)
@@ -106,7 +106,7 @@ struct pmc_md_ucf_pmc {
 };
 
 struct pmc_md_ucp_pmc {
-	uint32_t	pm_ucp_evsel;
+	uint64_t	pm_ucp_evsel;
 };
 
 /*
